+++ CMakeLists.txt
@@ -195,7 +195,7 @@ else ()
     add_subdirectory(src/data)
     add_subdirectory(src/edge)
     add_subdirectory(src/emulator)
-    add_subdirectory(src/http)
+    add_subdirectory(src/endpoint)
     add_subdirectory(src/knative)
     add_subdirectory(src/lambda)
     add_subdirectory(src/memory)
