+++ experiment/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 add_executable(http_experiment http_experiment.cpp)
 
-target_link_libraries(http_experiment pistache util)
+target_link_libraries(http_experiment pistache endpoint pthread util)
 
 add_executable(mem_experiment mem_experiment.cpp)
