+++ experiment/http_experiment.cpp
@@ -1,15 +1,28 @@
 #include "pistache/endpoint.h"
 #include <util/logging.h>
 #include <signal.h>
+#include <endpoint/Endpoint.h>
 
 using namespace Pistache;
 
+
 class HelloHandler : public Http::Handler {
 public:
-    HTTP_PROTOTYPE(HelloHandler)
+HTTP_PROTOTYPE(HelloHandler)
 
-    void onRequest(const Http::Request& request, Http::ResponseWriter response) override{
+    void onRequest(const Http::Request &request, Http::ResponseWriter response) override {
         response.send(Pistache::Http::Code::Ok, "Hello World\n");
+    }
+};
+
+class HelloEndpoint : public endpoint::Endpoint {
+public:
+    HelloEndpoint(int port, int threads): endpoint::Endpoint(port, threads) {
+
+    }
+
+    std::shared_ptr<Pistache::Http::Handler> getHandler() {
+        return Pistache::Http::make_handler<HelloHandler>();
     }
 };
 
@@ -27,14 +40,8 @@ int main() {
         throw std::runtime_error("Install signal handler failed");
     }
 
-    Pistache::Address addr(Pistache::Ipv4::any(), Pistache::Port(8080));
-    auto opts = Pistache::Http::Endpoint::options()
-            .threads(5);
-
-    Http::Endpoint server(addr);
-    server.init(opts);
-    server.setHandler(Http::make_handler<HelloHandler>());
-    server.serve();
+    HelloEndpoint server(8080, 5);
+    server.start();
 
     // Wait for a signal
     int signal = 0;
